FB Commit #4
HUGE UPDATE: today I've discovered that I can have blend files in the Godo engine and access their models and import them as scenes. this cleans up a lot of mess and shortens the pipeline for models/meshes. so currently I've purged all the terrain models and am going to move them to blend files